# .synnovator/openapi.yaml
openapi: 3.0.3
info:
  title: Synnovator API
  version: 1.0.0
  description: API for Synnovator - Activity and Competition Management Platform
  contact:
    name: API Support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
- url: http://localhost:8000
  description: Development server
- url: https://api.synnovator.com
  description: Production server
...
paths:
...
  /posts/{post_id}/like:
    post:
      summary: Like post
      operationId: like_post
      tags:
      - interactions
      parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '201':
          description: Post liked
        '409':
          description: Already liked
    delete:
      summary: Unlike post
      operationId: unlike_post
      tags:
      - interactions
      parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Like removed
  /posts/{post_id}/comments:
    get:
      summary: List post comments
      operationId: list_post_comments
      tags:
      - interactions
      parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: string
      - $ref: '#/components/parameters/SkipParam'
      - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: List of comments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCommentList'
    post:
      summary: Add comment
      operationId: add_post_comment
      tags:
      - interactions
      parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentCreate'
      responses:
        '201':
          description: Comment added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
...
  /posts/{post_id}/ratings:
    get:
      summary: List post ratings
      operationId: list_post_ratings
      tags:
      - interactions
      parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: string
      - $ref: '#/components/parameters/SkipParam'
      - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: List of ratings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedRatingList'
    post:
      summary: Submit rating
      operationId: submit_post_rating
      tags:
      - interactions
      parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RatingCreate'
      responses:
        '201':
          description: Rating submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rating'
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: /oauth/authorize
          tokenUrl: /oauth/token
          scopes:
            read: Read access to resources
            write: Write access to resources
            admin: Admin access for batch operations
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserRole:
      type: string
      enum:
      - participant
      - organizer
      - admin
      description: User role
...
    User:
      type: object
      required:
      - id
      - username
      - email
      - role
      - created_at
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        display_name:
          type: string
        avatar_url:
          type: string
          format: uri
        bio:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
...
    CommentCreate:
      type: object
      required:
      - content
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 2000
        parent_id:
          type: string
          description: Parent comment ID for replies
    Comment:
      type: object
      required:
      - id
      - content
      - created_by
      - created_at
      properties:
        id:
          type: string
          example: iact_abc123
        content:
          type: string
        parent_id:
          type: string
        created_by:
          type: string
        author:
          $ref: '#/components/schemas/User'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RatingCreate:
      type: object
      required:
      - scores
      properties:
        scores:
          type: object
          additionalProperties:
            type: number
            minimum: 0
            maximum: 100
          example:
            Innovation: 87
            Technical: 82
            Practical: 78
        comment:
          type: string
          description: Optional rating comment
    Rating:
      type: object
      required:
      - id
      - scores
      - created_by
      - created_at
      properties:
        id:
          type: string
        scores:
          type: object
          additionalProperties:
            type: number
        comment:
          type: string
        created_by:
          type: string
        author:
          $ref: '#/components/schemas/User'
        created_at:
          type: string
          format: date-time
...
    PaginatedCommentList:
      type: object
      required:
      - items
      - total
      - skip
      - limit
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
        total:
          type: integer
          description: Total number of records
        skip:
          type: integer
        limit:
          type: integer
    PaginatedRatingList:
      type: object
      required:
      - items
      - total
      - skip
      - limit
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Rating'
        total:
          type: integer
          description: Total number of records
        skip:
          type: integer
        limit:
          type: integer
...
  parameters:
    SkipParam:
      name: skip
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
      description: Number of records to skip
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      description: Maximum number of records to return
...
security:
- oauth2:
  - read
  - write